requirements.txt
python-dotenv
streamlit
langchain
langchain-core
langchain-groq
groq
Pillow
pydantic
#TTS